hard

+++ final_project/include/myfile.h
@@ -11,16 +11,16 @@ private:
   string file_name;
   int file_no;
   string file_pin;
-  int file_size;
+  long long int file_size;
 public:
   myfile();
-  myfile(string, int, string, int);
+  myfile(string, int, string, long long int);
   ~myfile();
 
   string get_file_name();
   int get_file_no();
   string get_file_pin();
-  int get_file_size();
+  long long int get_file_size();
 };
 
 #endif
